lambdas/alb-log-ingestor/requirements.txt
aws-requests-auth==0.4.1
elasticsearch==5.4.0
python-dateutil==2.6.0
